feature: rollup integration for UMD build
Integrated Rollup as a build tool to improve compatibility outside of the ES Module ecosystem. The tool is configured to build a UMD bundle that can be used as an ES Module, AMD module, CommonJS module or as a standalone script in the browser. This latter case assumes leaflet is loaded and the `L` variable is available in the global scope; the `TileLayerHeaders` class will be exposed under the property name `L.TileLayerHeaders`